combine BlockExecutor into ContractRuntime components - part 1
