+++ .gitignore
@@ -137,6 +137,9 @@ resources/local/chainspec.toml
 .idea/
 cmake-build-debug/
 
+# vscode
+.vscode/
+
 # NCTL transient assets
 /utils/nctl/assets
 /utils/nctl/dumps
