+++ node/src/components.rs
@@ -2,7 +2,6 @@
 //!
 //! Components are the building blocks of the whole application, wired together inside a reactor.
 //! Each component has a unified interface, expressed by the `Component` trait.
-pub(crate) mod block_executor;
 pub(crate) mod block_proposer;
 pub(crate) mod block_validator;
 pub(crate) mod chainspec_loader;
