+++ node/src/reactor/joiner.rs
@@ -29,7 +29,6 @@ use crate::components::{
 use crate::testing::network::NetworkedReactor;
 use crate::{
     components::{
-        block_executor::{self, BlockExecutor},
         block_validator::{self, BlockValidator},
         chainspec_loader::{self, ChainspecLoader},
         consensus::{self, EraId, EraSupervisor, HighwayProtocol},
@@ -49,15 +48,14 @@ use crate::{
     },
     effect::{
         announcements::{
-            BlockExecutorAnnouncement, ChainspecLoaderAnnouncement, ConsensusAnnouncement,
+            ChainspecLoaderAnnouncement, ConsensusAnnouncement, ContractRuntimeAnnouncement,
             DeployAcceptorAnnouncement, GossiperAnnouncement, LinearChainAnnouncement,
             NetworkAnnouncement,
         },
         requests::{
-            BlockExecutorRequest, BlockProposerRequest, BlockValidationRequest,
-            ChainspecLoaderRequest, ConsensusRequest, ContractRuntimeRequest, FetcherRequest,
-            LinearChainRequest, MetricsRequest, NetworkInfoRequest, NetworkRequest, RestRequest,
-            StateStoreRequest, StorageRequest,
+            BlockProposerRequest, BlockValidationRequest, ChainspecLoaderRequest, ConsensusRequest,
+            ContractRuntimeRequest, FetcherRequest, LinearChainRequest, MetricsRequest,
+            NetworkInfoRequest, NetworkRequest, RestRequest, StateStoreRequest, StorageRequest,
         },
         EffectBuilder, Effects,
     },
@@ -139,10 +137,6 @@ pub enum Event {
     #[from]
     LinearChainSync(#[serde(skip_serializing)] linear_chain_sync::Event<NodeId>),
 
-    /// Block executor event.
-    #[from]
-    BlockExecutor(#[serde(skip_serializing)] block_executor::Event),
-
     /// Contract Runtime event.
     #[from]
     ContractRuntime(#[serde(skip_serializing)] contract_runtime::Event),
@@ -176,10 +170,6 @@ pub enum Event {
     #[from]
     BlockValidatorRequest(#[serde(skip_serializing)] BlockValidationRequest<Block, NodeId>),
 
-    /// Block executor request.
-    #[from]
-    BlockExecutorRequest(#[serde(skip_serializing)] BlockExecutorRequest),
-
     /// Block proposer request.
     #[from]
     BlockProposerRequest(#[serde(skip_serializing)] BlockProposerRequest),
@@ -201,7 +191,7 @@ pub enum Event {
 
     /// Block executor announcement.
     #[from]
-    BlockExecutorAnnouncement(#[serde(skip_serializing)] BlockExecutorAnnouncement),
+    ContractRuntimeAnnouncement(#[serde(skip_serializing)] ContractRuntimeAnnouncement),
 
     /// Consensus announcement.
     #[from]
@@ -299,14 +289,10 @@ impl Display for Event {
             }
             Event::BlockValidator(event) => write!(f, "block validator event: {}", event),
             Event::DeployFetcher(event) => write!(f, "deploy fetcher event: {}", event),
-            Event::BlockExecutor(event) => write!(f, "block executor event: {}", event),
-            Event::BlockExecutorRequest(request) => {
-                write!(f, "block executor request: {}", request)
-            }
             Event::BlockProposerRequest(req) => write!(f, "block proposer request: {}", req),
-            Event::ContractRuntime(event) => write!(f, "contract runtime event: {}", event),
+            Event::ContractRuntime(event) => write!(f, "contract runtime event: {:?}", event),
             Event::LinearChain(event) => write!(f, "linear chain event: {}", event),
-            Event::BlockExecutorAnnouncement(announcement) => {
+            Event::ContractRuntimeAnnouncement(announcement) => {
                 write!(f, "block executor announcement: {}", announcement)
             }
             Event::Consensus(event) => write!(f, "consensus event: {}", event),
@@ -347,7 +333,6 @@ pub struct Reactor {
     pub(super) linear_chain_sync: LinearChainSync<NodeId>,
     pub(super) block_validator: BlockValidator<Block, NodeId>,
     pub(super) deploy_fetcher: Fetcher<Deploy>,
-    pub(super) block_executor: BlockExecutor,
     pub(super) linear_chain: linear_chain::LinearChain<NodeId>,
     pub(super) consensus: EraSupervisor<NodeId>,
     // Handles request for linear chain block by height.
@@ -385,7 +370,7 @@ impl reactor::Reactor for Reactor {
             config,
             chainspec_loader,
             storage,
-            contract_runtime,
+            mut contract_runtime,
             small_network_identity,
             network_identity,
         } = initializer;
@@ -475,11 +460,7 @@ impl reactor::Reactor for Reactor {
             DeployAcceptor::new(config.deploy_acceptor, &*chainspec_loader.chainspec());
 
         let genesis_state_root_hash = chainspec_loader.genesis_state_root_hash();
-        let block_executor = BlockExecutor::new(
-            genesis_state_root_hash,
-            protocol_version.clone(),
-            registry.clone(),
-        );
+        contract_runtime.set_genesis_state_root_hash(genesis_state_root_hash);
 
         let linear_chain = linear_chain::LinearChain::new(&registry)?;
 
@@ -537,7 +518,6 @@ impl reactor::Reactor for Reactor {
                 linear_chain_fetcher,
                 block_validator,
                 deploy_fetcher,
-                block_executor,
                 linear_chain,
                 consensus,
                 block_by_height_fetcher,
@@ -727,19 +707,12 @@ impl reactor::Reactor for Reactor {
                 rng,
                 Event::BlockByHeightFetcher(request.into()),
             ),
-            Event::BlockExecutor(event) => reactor::wrap_effects(
-                Event::BlockExecutor,
-                self.block_executor.handle_event(effect_builder, rng, event),
-            ),
-            Event::BlockExecutorRequest(request) => {
-                self.dispatch_event(effect_builder, rng, Event::BlockExecutor(request.into()))
-            }
             Event::ContractRuntime(event) => reactor::wrap_effects(
                 Event::ContractRuntime,
                 self.contract_runtime
                     .handle_event(effect_builder, rng, event),
             ),
-            Event::BlockExecutorAnnouncement(BlockExecutorAnnouncement::LinearChainBlock {
+            Event::ContractRuntimeAnnouncement(ContractRuntimeAnnouncement::LinearChainBlock {
                 block,
                 execution_results,
             }) => {
