+++ node/src/reactor/validator/memory_metrics.rs
@@ -40,8 +40,6 @@ pub(super) struct MemoryMetrics {
     mem_deploy_gossiper: IntGauge,
     /// Estimated heap memory usage of block_proposer component.
     mem_block_proposer: IntGauge,
-    /// Estimated heap memory usage of block executor component.
-    mem_block_executor: IntGauge,
     /// Estimated heap memory usage of block validator component.
     mem_proto_block_validator: IntGauge,
     /// Estimated heap memory usage of linear chain component.
@@ -89,8 +87,6 @@ impl MemoryMetrics {
         )?;
         let mem_block_proposer =
             IntGauge::new("mem_block_proposer", "block_proposer memory usage in bytes")?;
-        let mem_block_executor =
-            IntGauge::new("mem_block_executor", "block_executor memory usage in bytes")?;
         let mem_proto_block_validator = IntGauge::new(
             "mem_proto_block_validator",
             "proto_block_validator memory usage in bytes",
@@ -121,7 +117,6 @@ impl MemoryMetrics {
         registry.register(Box::new(mem_deploy_fetcher.clone()))?;
         registry.register(Box::new(mem_deploy_gossiper.clone()))?;
         registry.register(Box::new(mem_block_proposer.clone()))?;
-        registry.register(Box::new(mem_block_executor.clone()))?;
         registry.register(Box::new(mem_proto_block_validator.clone()))?;
         registry.register(Box::new(mem_linear_chain.clone()))?;
         registry.register(Box::new(mem_estimator_runtime_s.clone()))?;
@@ -141,7 +136,6 @@ impl MemoryMetrics {
             mem_deploy_fetcher,
             mem_deploy_gossiper,
             mem_block_proposer,
-            mem_block_executor,
             mem_proto_block_validator,
             mem_linear_chain,
             mem_estimator_runtime_s,
@@ -170,7 +164,6 @@ impl MemoryMetrics {
         let deploy_fetcher = reactor.deploy_fetcher.estimate_heap_size() as i64;
         let deploy_gossiper = reactor.deploy_gossiper.estimate_heap_size() as i64;
         let block_proposer = reactor.block_proposer.estimate_heap_size() as i64;
-        let block_executor = reactor.block_executor.estimate_heap_size() as i64;
         let proto_block_validator = reactor.proto_block_validator.estimate_heap_size() as i64;
 
         let linear_chain = reactor.linear_chain.estimate_heap_size() as i64;
@@ -188,7 +181,6 @@ impl MemoryMetrics {
             + deploy_fetcher
             + deploy_gossiper
             + block_proposer
-            + block_executor
             + proto_block_validator
             + linear_chain;
 
@@ -206,7 +198,6 @@ impl MemoryMetrics {
         self.mem_deploy_fetcher.set(deploy_fetcher);
         self.mem_deploy_gossiper.set(deploy_gossiper);
         self.mem_block_proposer.set(block_proposer);
-        self.mem_block_executor.set(block_executor);
         self.mem_proto_block_validator.set(proto_block_validator);
         self.mem_linear_chain.set(linear_chain);
 
@@ -228,7 +219,6 @@ impl MemoryMetrics {
                %deploy_fetcher,
                %deploy_gossiper,
                %block_proposer,
-               %block_executor,
                %proto_block_validator,
                %linear_chain,
                "Collected new set of memory metrics.");
@@ -296,11 +286,6 @@ impl Drop for MemoryMetrics {
             .unregister(Box::new(self.mem_block_proposer.clone()))
             .unwrap_or_else(
                 |err| warn!(%err, "did not expect deregistering mem_block_proposer, to fail"),
-            );
-        self.registry
-            .unregister(Box::new(self.mem_block_executor.clone()))
-            .unwrap_or_else(
-                |err| warn!(%err, "did not expect deregistering mem_block_executor, to fail"),
             );
         self.registry
             .unregister(Box::new(self.mem_proto_block_validator.clone()))
